+++ resources/lang/en/reporting.php
@@ -7,5 +7,5 @@ return [
     'kitchen' => 'Kitchen',
     'user-privileges' => 'User privileges',
     'role-privileges' => 'Role privileges',
-    'user-access-sensitive-data' => 'Users with access to sensitive data',
+    'privacy' => 'Privacy report',
 ];