+++ Changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.12.5
+
+* UI fixes
+
+## 1.12.4
+
+* Show administrators in pricacy report
+* Logviewer table improvements
+* Added view about latest donations
+
 ## 1.12.3
 
 * Adder user privilege repots to reporting screen